Update README.md
I really had Problems setting it up, but a friend of mine who is more experienced was able to help me. I would have never found out, that I need to Install pkg-config, as it was not clear from the error message. 

I hope this is alright. I think somewhere should be the hint, I am NotSure though if it is the right place.

Really great Software. Thank you it saved my build!